+++ core/src/test/java/io/xj/core/dao/PhaseChordIT.java
@@ -77,14 +77,14 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setName("G minor 7")
       .setPhaseId(BigInteger.valueOf(2));
 
     JSONObject result = JSON.objectFrom(testDAO.create(access, inputData));
 
     assertNotNull(result);
-    assertEquals(4, result.get("position"));
+    assertEquals(4.0, result.get("position"));
     assertEquals("G minor 7", result.get("name"));
     assertEquals(2, result.get("phaseId"));
   }
@@ -96,7 +96,7 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setName("G minor 7");
 
     testDAO.create(access, inputData);
@@ -109,7 +109,7 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setPhaseId(BigInteger.valueOf(2));
 
     testDAO.create(access, inputData);
@@ -179,7 +179,7 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setName("G minor 7");
 
     testDAO.update(access, BigInteger.valueOf(3), inputData);
@@ -192,7 +192,7 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setPhaseId(BigInteger.valueOf(2));
 
     testDAO.update(access, BigInteger.valueOf(2), inputData);
@@ -205,7 +205,7 @@ public class PhaseChordIT {
       "accounts", "1"
     ));
     PhaseChord inputData = new PhaseChord()
-      .setPosition(4)
+      .setPosition(4.0)
       .setPhaseId(BigInteger.valueOf(57))
       .setName("D minor");
 
@@ -230,14 +230,14 @@ public class PhaseChordIT {
     PhaseChord inputData = new PhaseChord()
       .setPhaseId(BigInteger.valueOf(1))
       .setName("POPPYCOCK")
-      .setPosition(4);
+      .setPosition(4.0);
 
     testDAO.update(access, BigInteger.valueOf(1), inputData);
 
     PhaseChord result = testDAO.readOne(Access.internal(), BigInteger.valueOf(1));
     assertNotNull(result);
     assertEquals("POPPYCOCK", result.getName());
-    assertEquals((Integer) 4, result.getPosition());
+    assertEquals(Double.valueOf(4), result.getPosition());
     assertEquals(BigInteger.valueOf(1), result.getPhaseId());
   }
